qemu.spec: Add dependency on seabios 1.10.2-4
The seabios that comes with OL 7.4 has version string 1.10.2-3. The
seabios-bin rpm built from the ca-git repo has patches to fix a bug
reported in the seabios that comes with OL 7.4. The seabios built from
the ca-git repo has version string 1.10.2-4 and to force QEMU to
pull this rpm, a version dependency needs to be added to the QEMU
spec file.

Orabug: 26755701

Signed-off-by: Ameya More <ameya.more@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>