virtio_net: Add support for "Data Path Switching" during Live Migration.
Added two new events, FAILOVER_PLUG_PRIMARY and FAILOVER_UNPLUG_PRIMARY.
The first is emitted when the guest negotiates the F_STANDBY feature
bit. The second is emitted when the virtio_net driver is removed, either
manually or as a result of guest reboot.

Also, the Virtual Functions will be automatically removed from the guest
if the guest is rebooted. To properly identify the VFIO devices that
must be removed, a new property named "x-failover-primary" is added to
the vfio-pci devices. Only the vfio-pci devices that have this property
enabled are removed from the guest upon reboot.

Orabug: 28732905

Signed-off-by: Venu Busireddy <venu.busireddy@oracle.com>
Reviewed-by: Si-Wei Liu <si-wei.liu@oracle.com>