vfio-pci: add FAILOVER_PRIMARY_CHANGED event to shorten downtime during failover
When hot plugging in a VF, datapath switching will be performed immediately which
is sub-optimal in terms of timing, and could end up with substantial downtime.
One of ways to shorten this downtime is to switch datapath until after VF is seen
to get enabled by guest. That is when bus master bit in VF's PCI config space
gets enabled, a corresponding FAILOVER_PRIMARY_CHANGED event will be emitted to
indicate this condition. Then management stack can kick off datapath switching
upon receiving the event.

Orabug: 28833460

Signed-off-by: Si-Wei Liu <si-wei.liu@oracle.com>
Reviewed-by: Venu Busireddy <venu.busireddy@oracle.com>