parfait: provide option to upload results
By default when running 'rpmbuild --with parfait' the comparison
is run against a baseline (using the -b option and a URL based
on %version).  It is desirable to provide a way to specify that
results should be uploaded to a Parfait Server.

This patch adds support so that if '--define "_parfait_upload 1"'
is passed to rpmbuild, the results are uploaded to the Parfait
Server specified by '_parfait_server'.  When '_parfait_upload' is
set, the the source root directory is also set via -z to shorten
reported filenames.

Orabug: 28733157

Signed-off-by: Liam Merwick <liam.merwick@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>
(cherry picked from commit abc20a25b0138274448909823ff9845ecd591cc9)
Signed-off-by: Lidong Chen <lidong.chen@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>