parfait: deal with parfait returning non-zero return value
Now that the Parfait tooling bug, 21116133, is fixed in Parfait 2.3.5,
some QEMU object files no longer trigger internal parfait tool failures
so the spec file no longer needs to check for exit code 3.

Work is ongoing to triage all the issues reported by Parfait on the QEMU
codebase but until they are all are either fixed or marked as a
False Positive/Will-Not-Fix on the QEMU Parfait server ($_parfait_server),
the parfait run will return 1 so we still need to check the return value
and ignore it if it's 1 or (some time in the future) zero (at which stage
this check can be removed).

Note: The existing code was broken in that it always returned success
irrespective of the parfait exit code.

Orabug: 28733157

Signed-off-by: Liam Merwick <liam.merwick@oracle.com>
Reviewed-by: Darren Kenny <Darren.Kenny@oracle.com>
(cherry picked from commit 7dc85b33bf444eefd69e7744bb1004959e42f702)
Signed-off-by: Lidong Chen <lidong.chen@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>