parfait: --disable-avx2 no longer needed by rpmbuild
The --disable-avx2 option is no longer needed to compile the QEMU
codebase with Parfait so we should remove it from buildrpm/qemu.spec
It is not clear when the dependency changed - trying a number of older
Parfait releases back to 2.0.0, we now don't hit the AVX2 issue and
neither do older releases of QEMU so it may be due to upgraded build tools.

The impetus for this change is for the backport of Parfait support
to 2.9 which won't compile with --disable-avx2.

Orabug: 28733157

Signed-off-by: Liam Merwick <liam.merwick@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
(cherry picked from commit d9e750075ebbc6954b6dd7ee1b1bee75ff1c8507)
Signed-off-by: Lidong Chen <lidong.chen@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>