migration: fix max_cpu_throttle lower limit check
In QEMU 2.9.0, when setting the 'max_cpu_throttle' migration parameter, the
code does not correctly check 'cpu_throttle_initial' as the the lower limit.
The issue is 'params' in qmp_migrate_set_parameters() refers only to newly set
parameter(s), and not the existing settings. The fix is to check the existing
'cpu_throttle_initial' value for the lower limit.

Fixes: e6d7378070f2 ("migrate/cpu-throttle: Add max-cpu-throttle migration parameter")

Orabug: 30149517

Reviewed-by: Liam Merwick <liam.merwick@oracle.com>
Signed-off-by: Mark Kanda <mark.kanda@oracle.com>